+++ internal/store/deployment_test.go
@@ -55,6 +55,8 @@ func TestDeploymentStore(t *testing.T) {
 		# TYPE kube_deployment_spec_replicas gauge
 		# HELP kube_deployment_status_replicas The number of replicas per deployment.
 		# TYPE kube_deployment_status_replicas gauge
+		# HELP kube_deployment_status_replicas_ready The number of ready replicas per deployment.
+		# TYPE kube_deployment_status_replicas_ready gauge
 		# HELP kube_deployment_status_replicas_available The number of available replicas per deployment.
 		# TYPE kube_deployment_status_replicas_available gauge
 		# HELP kube_deployment_status_replicas_unavailable The number of unavailable replicas per deployment.
@@ -90,6 +92,7 @@ func TestDeploymentStore(t *testing.T) {
 				},
 				Status: v1.DeploymentStatus{
 					Replicas:            15,
+					ReadyReplicas:       10,
 					AvailableReplicas:   10,
 					UnavailableReplicas: 5,
 					UpdatedReplicas:     2,
@@ -123,6 +126,7 @@ func TestDeploymentStore(t *testing.T) {
         kube_deployment_status_replicas_unavailable{deployment="depl1",namespace="ns1"} 5
         kube_deployment_status_replicas_updated{deployment="depl1",namespace="ns1"} 2
         kube_deployment_status_replicas{deployment="depl1",namespace="ns1"} 15
+        kube_deployment_status_replicas_ready{deployment="depl1",namespace="ns1"} 10
         kube_deployment_status_condition{deployment="depl1",namespace="ns1",condition="Available",status="true"} 1
         kube_deployment_status_condition{deployment="depl1",namespace="ns1",condition="Progressing",status="true"} 1
         kube_deployment_status_condition{deployment="depl1",namespace="ns1",condition="Available",status="false"} 0
@@ -143,6 +147,7 @@ func TestDeploymentStore(t *testing.T) {
 				},
 				Status: v1.DeploymentStatus{
 					Replicas:            10,
+					ReadyReplicas:       5,
 					AvailableReplicas:   5,
 					UnavailableReplicas: 0,
 					UpdatedReplicas:     1,
@@ -177,6 +182,7 @@ func TestDeploymentStore(t *testing.T) {
         kube_deployment_status_replicas_unavailable{deployment="depl2",namespace="ns2"} 0
         kube_deployment_status_replicas_updated{deployment="depl2",namespace="ns2"} 1
         kube_deployment_status_replicas{deployment="depl2",namespace="ns2"} 10
+        kube_deployment_status_replicas_ready{deployment="depl2",namespace="ns2"} 5
         kube_deployment_status_condition{deployment="depl2",namespace="ns2",condition="Available",status="true"} 0
         kube_deployment_status_condition{deployment="depl2",namespace="ns2",condition="Progressing",status="true"} 0
         kube_deployment_status_condition{deployment="depl2",namespace="ns2",condition="ReplicaFailure",status="true"} 1