+++ internal/store/node.go
@@ -93,6 +93,7 @@ func createNodeInfoFamilyGenerator() generator.FamilyGenerator {
 				"kubeproxy_version",
 				"provider_id",
 				"pod_cidr",
+				"system_uuid",
 			}
 			labelValues := []string{
 				n.Status.NodeInfo.KernelVersion,
@@ -102,6 +103,7 @@ func createNodeInfoFamilyGenerator() generator.FamilyGenerator {
 				n.Status.NodeInfo.KubeProxyVersion,
 				n.Spec.ProviderID,
 				n.Spec.PodCIDR,
+				n.Status.NodeInfo.SystemUUID,
 			}
 
 			internalIP := ""